+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- [PR [#64](https://github.com/DmitryTsepelev/ar_lazy_preload/pull/64)] Create context after `find_by`/`find_by!` when `auto_preload` is on ([@DmitryTsepelev][])
+
 ## 1.1.1 (2022-12-01 🤦‍♂️)
 
 - [PR [#63](https://github.com/DmitryTsepelev/ar_lazy_preload/pull/63)] Pass arguments to `.reload` ([@DmitryTsepelev][])